Test/requirements.txt.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dotenv
playwright
pandas
beautifulsoup4
openai
pdfplumber
